Update Battle scene

+++ Project-Pentagram/Game/BattleScene/BattleScene.h
@@ -14,7 +14,7 @@ private:
 public:
     virtual void GameSceneLoad() override;
     virtual void GameSceneInit() override;
-    virtual void GameSceneUpdate(double dt) override;
+    virtual void GameSceneUpdate(float dt) override;
     virtual void GameSceneDraw() override;
     virtual void GameSceneUnload() override;
     virtual void GameSceneFree() override;
